+++ src/main/java/janggi/dao/entity/GameEntity.java
@@ -1,20 +1,20 @@
-package janggi.dao;
+package janggi.dao.entity;
 
 import janggi.domain.piece.Dynasty;
 
-public class Game {
+public class GameEntity {
 
     private Long id;
     private Status status;
     private Dynasty currentTurn;
 
-    public Game(Long id, Status status, Dynasty currentTurn) {
+    public GameEntity(Long id, Status status, Dynasty currentTurn) {
         this.id = id;
         this.status = status;
         this.currentTurn = currentTurn;
     }
 
-    public Game(Status status, Dynasty currentTurn) {
+    public GameEntity(Status status, Dynasty currentTurn) {
         this.status = status;
         this.currentTurn = currentTurn;
     }
