+++ src/main/java/janggi/dao/entity/Status.java
@@ -1,4 +1,4 @@
-package janggi.dao;
+package janggi.dao.entity;
 
 import java.util.Arrays;
